+++ inference.py
@@ -1,55 +1,68 @@
 import torch
-import torch.nn.functional as F
+import torchvision.transforms as transforms
+import clip
+from datasets import build_dataset
+from datasets.utils import build_data_loader
+from PIL import Image
+
 
 from utils import *
+from run_utils import *
+from lora import run_lora_inference, lora_inference
 
-from loralib.utils import apply_lora, load_lora
-from loralib import layers as lora_layers
 
+def inference(args, image):
 
-def lora_inference(args, clip_model, image, dataset):
-    """
-    Dado um tensor de imagem, retorna o template que gera maior similaridade para essa imagem.
-    """
+    
+    set_random_seed(args.seed)
+    
+    # CLIP
+    clip_model, preprocess = clip.load(args.backbone)
     clip_model.eval()
-    best_template = None
-    best_similarity = -float('inf')
-    best_idx = -1
-    with torch.no_grad():
-        # Normaliza a imagem e move para cuda
-        image = image.cuda()
-        with torch.amp.autocast(device_type="cuda", dtype=torch.float16):
-            image_features = clip_model.encode_image(image)
-        image_features = image_features / image_features.norm(dim=-1, keepdim=True)
+    logit_scale = 100
 
-        # Testa todos os templates
-        for idx, template in enumerate(dataset.template):
-            texts = [template.format(classname.replace('_', ' ')) for classname in dataset.classnames]
-            with torch.amp.autocast(device_type="cuda", dtype=torch.float16):
-                texts_tokenized = clip.tokenize(texts).cuda()
-                class_embeddings = clip_model.encode_text(texts_tokenized)
-            text_features = class_embeddings / class_embeddings.norm(dim=-1, keepdim=True)
-            # Similaridade da imagem com todas as classes desse template
-            cosine_similarity = (image_features @ text_features.t()).squeeze(0)
-            max_sim = cosine_similarity.max().item()
-            if max_sim > best_similarity:
-                best_similarity = max_sim
-                best_template = template
-                best_idx = idx
-    return best_template
+    # Prepare dataset
+    print("Preparing dataset.")
+        
+    dataset = build_dataset(args.dataset, args.root_path, args.shots, preprocess)
+    
+
+    template = run_lora_inference(args, clip_model, image, dataset)
+
+    print(f"Best template for the image: {template}")
 
 
-def run_lora_inference(args, clip_model, image, dataset):
-    """
-    Recebe uma imagem, roda o lora_inference e exibe o melhor template para essa imagem.
-    """
-    list_lora_layers = apply_lora(args, clip_model)
-    clip_model = clip_model.cuda()
-    load_lora(args, list_lora_layers)
-    best_template = lora_inference(args, clip_model, image, dataset)
-    print(f"Melhor template para a imagem: {best_template}")
-    return best_template
+args = get_arguments()
+
+img_path = args.image_path
+image = Image.open(img_path).convert("RGB")
+
+# Use o mesmo preprocess do CLIP
+preprocess = transforms.Compose([
+            transforms.RandomResizedCrop(size=224, scale=(0.08, 1), interpolation=transforms.InterpolationMode.BICUBIC),
+            transforms.RandomHorizontalFlip(p=0.5),
+            transforms.ToTensor(),
+            transforms.Normalize(mean=(0.48145466, 0.4578275, 0.40821073), std=(0.26862954, 0.26130258, 0.27577711))
+        ])
+
+image_tensor = preprocess(image).unsqueeze(0)  # Adiciona batch dimension
+
+# Rode a inferência
+inference(args, image_tensor)
+
+'''
+to run inference:
+
+python inference.py \
+  --root_path /home/pedro36/projects/def-leszek/pedro36/datasets/DATA \
+  --dataset pigs \
+  --seed 1 \
+  --shots 1
+  --save_path weights \
+  --image_path "/home/pedro36/projects/def-leszek/pedro36/workspace/CLIP-LoRA-Original/cropped_image_6359.jpg" \
+  --filename "CLIP-LoRA_pigs"
 
 
-
+python inference.py --root_path /home/pedro36/projects/def-leszek/pedro36/datasets/DATA --dataset pigs --seed 1 --shots 1 --save_path weights --image_path "/home/pedro36/projects/def-leszek/pedro36/workspace/CLIP-LoRA-Original/cropped_image_6359.jpg" --filename "CLIP-LoRA_pigs"
+'''
